photo-animation/requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.5.0
opencv-python>=4.7.0
numpy>=1.24.0
safetensors>=0.3.0
